+++ src/inputrules/ed-keymap.js
@@ -1,0 +1,7 @@
+import {buildKeymap} from 'prosemirror/dist/example-setup'
+import EdSchema from '../schema/ed-schema-full'
+
+
+const EdKeymap = buildKeymap(EdSchema)
+
+export default EdKeymap