+++ src/store/ed-store.js
@@ -26,12 +26,14 @@ export default class EdStore {
     this.on('change', options.onChange)
     options.onChange = this.routeChange.bind(this)
     this.onShareUrl = options.onShareUrl
-    this.onShareFile = options.onShareFile
     this.onPlaceholderCancel = options.onPlaceholderCancel || noop
     this.onCommandsChanged = options.onCommandsChanged
     this.onRequestCoverUpload = options.onRequestCoverUpload
     options.onDropFiles = options.onDropFiles || noop
     this.onDropFileOnBlock = options.onDropFileOnBlock || noop
+
+    this.onShareFile = options.onShareFile || noop
+    this.on('command.menu.file', this.onShareFile)
 
     // Listen for first render
     this.on('plugin.widget.initialized', options.onMount || noop)
@@ -195,9 +197,6 @@ export default class EdStore {
     this.pm.tr
       .delete(pos, pos + nodeToRemove.nodeSize)
       .apply()
-
-    // Trigger event for widget system
-    setTimeout(() => this.pm.signal('draw'), 0)
   }
   _dedupeIds () {
     let ids = []
@@ -216,8 +215,6 @@ export default class EdStore {
       }
       ids.push(id)
     }
-    // Trigger event for widget system
-    setTimeout(() => this.pm.signal('draw'), 0)
   }
   getBlock (id) {
     return this._content[id]
@@ -246,9 +243,6 @@ export default class EdStore {
       // Insert the block
       .insert(pos, node)
       .apply()
-
-    // Trigger event for widget system
-    setTimeout(() => this.pm.signal('draw'), 0)
   }
   _insertBlocks (index, blocks) {
     if (!this.pm) {
@@ -268,9 +262,6 @@ export default class EdStore {
       const pos = indexToPos(this.pm.doc, index + i)
       this.pm.tr.insert(pos, node).apply()
     }
-
-    // Trigger event for widget system
-    setTimeout(() => this.pm.signal('draw'), 0)
   }
   insertPlaceholders (index, count) {
     let toInsert = []
@@ -374,8 +365,6 @@ export default class EdStore {
         .apply()
     }
 
-    // Trigger event for widget system
-    setTimeout(() => this.pm.signal('draw'), 0)
     // Focus first textblock
     try {
       this.pm.checkPos(1, true)
@@ -385,17 +374,13 @@ export default class EdStore {
     this.pm.scrollIntoView()
   }
   getContent () {
-    const doc = this.pm.getContent()
-    const content = DocToGrid(doc, this._content)
-    return content
+    return DocToGrid(this.pm.doc, this._content)
   }
   setContent (content) {
     this._applyTransform(content)
     this._initializeContent(content)
     // Let widgets know to update
     this.trigger('media.update')
-    // Trigger event for widget system
-    setTimeout(() => this.pm.signal('draw'), 0)
   }
   _applyTransform (content) {
     for (let i = 0, len = content.length; i < len; i++) {