+++ webpack.config.js
@@ -17,8 +17,7 @@ var loaders = [
     loader: 'babel-loader', 
     include: [
       path.resolve(__dirname, 'demo'),
-      path.resolve(__dirname, 'src'),
-      path.resolve(__dirname, 'node_modules', 'prosemirror')
+      path.resolve(__dirname, 'src')
     ]
   },
   { test: /\.css$/, loader: 'style?singleton!raw' },
@@ -28,7 +27,7 @@ var loaders = [
 if (__DEV || __DEMO) {
   entry.demo = './demo/demo.js'
 } else {
-  entry.ed = './src/ed.js'
+  entry.build = './src/ed.js'
 }
 
 if (__KARMA) {